% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{.k_means_fit_sparklyr}
\alias{.k_means_fit_sparklyr}
\title{Simple Wrapper around sparklyr ml_kmeans}
\usage{
.k_means_fit_sparklyr(
  data,
  formula = NULL,
  clusters,
  num_init = 1,
  max_iters = 20,
  initializer = "k-means||",
  CENTROIDS = NULL,
  tol = 1e-04,
  tol_optimal_init = 0.3,
  seed = NULL
)
}
\arguments{
\item{formula}{Used when \code{x} is a \code{tbl_spark}. R formula as a character string or a formula. This is used to transform the input dataframe before fitting, see \link[sparklyr]{ft_r_formula} for details.}

\item{tol}{Param for the convergence tolerance for iterative algorithms.}

\item{seed}{A random seed. Set this value if you need your results to be
reproducible across repeated calls.}

\item{...}{Other arguments passed to \code{sparklyr::ml_kmeans()}}
}
\value{
Result from \code{sparklyr::ml_kmeans()}
}
\description{
This wrapper runs \code{sparklyr::ml_kmeans()} and adds a check that \code{centers} is
specified. And reorders the clusters.
}
\keyword{internal}
